+++ lib/core/constants/api.dart
@@ -41,3 +41,6 @@ const String kAPICustomerProfileURL = '/customer/profile'; // GET, PUT
 
 //! category-controller
 const String kAPIGetAllCategoryURL = '/category/all-parent'; // GET
+
+//! product-suggestion-controller
+const String kAPIGetSuggestionProductURL = '/product-suggestion/get-page/randomly'; // GET